refactor(utils): reorganize package by splitting overloaded modules
Move number functions (validateOffset, parseNumericId, parseBigIntId,
formatPosition) from string.ts to number.ts. Consolidate all time-period
aggregation into aggregation.ts, eliminating duplication between
date-ranges.ts and score-aggregation.ts. Move chart-building functions
(buildChartData, formatLabel) from date-ranges.ts to chart.ts. Move
TTS_VOICES/TTSVoice from constants.ts to languages.ts.
