+++ apps/admin/src/app/(private)/review/[group]/[task]/unflag-action.ts
@@ -4,7 +4,7 @@ import { assertAdmin } from "@/lib/admin-guard";
 import { prisma } from "@zoonk/db";
 import { safeAsync } from "@zoonk/utils/error";
 import { parseFormField } from "@zoonk/utils/form";
-import { parseBigIntId } from "@zoonk/utils/string";
+import { parseBigIntId } from "@zoonk/utils/number";
 import { revalidatePath } from "next/cache";
 
 export async function unflagAction(formData: FormData) {
