+++ apps/admin/src/app/(private)/users/[id]/_actions/unban-user.ts
@@ -4,7 +4,7 @@ import { assertAdmin } from "@/lib/admin-guard";
 import { prisma } from "@zoonk/db";
 import { safeAsync } from "@zoonk/utils/error";
 import { parseFormField } from "@zoonk/utils/form";
-import { parseNumericId } from "@zoonk/utils/string";
+import { parseNumericId } from "@zoonk/utils/number";
 import { revalidatePath } from "next/cache";
 
 export async function unbanUserAction(formData: FormData) {
