+++ apps/admin/src/app/(private)/users/[id]/page.tsx
@@ -13,7 +13,7 @@ import {
   ContainerHeaderGroup,
 } from "@zoonk/ui/components/container";
 import { Skeleton } from "@zoonk/ui/components/skeleton";
-import { parseNumericId } from "@zoonk/utils/string";
+import { parseNumericId } from "@zoonk/utils/number";
 import Link from "next/link";
 import { notFound } from "next/navigation";
 import { Suspense } from "react";
