+++ apps/main/src/app/(catalog)/b/[brandSlug]/c/[courseSlug]/ch/[chapterSlug]/l/[lessonSlug]/lesson-header.tsx
@@ -19,7 +19,7 @@ import {
   MediaCardTitle,
   MediaCardTrigger,
 } from "@zoonk/ui/components/media-card";
-import { formatPosition } from "@zoonk/utils/string";
+import { formatPosition } from "@zoonk/utils/number";
 import { getExtracted } from "next-intl/server";
 import Link from "next/link";
