+++ apps/main/src/app/generate/cs/[id]/generate-course-suggestion-content.tsx
@@ -11,7 +11,8 @@ import {
 import { Empty, EmptyContent, EmptyHeader } from "@zoonk/ui/components/empty";
 import { Skeleton } from "@zoonk/ui/components/skeleton";
 import { AI_ORG_SLUG } from "@zoonk/utils/constants";
-import { ensureLocaleSuffix, parseNumericId } from "@zoonk/utils/string";
+import { parseNumericId } from "@zoonk/utils/number";
+import { ensureLocaleSuffix } from "@zoonk/utils/string";
 import { notFound, redirect } from "next/navigation";
 import { GenerationClient } from "./generation-client";
