+++ apps/main/src/data/progress/get-best-day.ts
@@ -1,8 +1,8 @@
 import "server-only";
 import { getSession } from "@zoonk/core/users/session/get";
 import { prisma } from "@zoonk/db";
+import { type ScoredRow, findBestByScore } from "@zoonk/utils/aggregation";
 import { getDefaultStartDate } from "@zoonk/utils/date-ranges";
-import { type ScoredRow, findBestByScore } from "@zoonk/utils/score-aggregation";
 import { cache } from "react";
 
 type BestDayData = {
