+++ apps/main/src/data/progress/get-bp-history.ts
@@ -1,15 +1,10 @@
 import "server-only";
 import { getSession } from "@zoonk/core/users/session/get";
 import { prisma } from "@zoonk/db";
+import { aggregateByMonth, aggregateByWeek, aggregateByYear } from "@zoonk/utils/aggregation";
 import { type BeltLevelResult, calculateBeltLevel } from "@zoonk/utils/belt-level";
-import {
-  type HistoryPeriod,
-  aggregateByMonth,
-  aggregateByWeek,
-  aggregateByYear,
-  calculateDateRanges,
-  formatLabel,
-} from "@zoonk/utils/date-ranges";
+import { formatLabel } from "@zoonk/utils/chart";
+import { type HistoryPeriod, calculateDateRanges } from "@zoonk/utils/date-ranges";
 import { safeAsync } from "@zoonk/utils/error";
 import { cache } from "react";
