+++ apps/main/src/data/progress/get-score-history.ts
@@ -1,13 +1,14 @@
 import "server-only";
 import { getSession } from "@zoonk/core/users/session/get";
 import { prisma } from "@zoonk/db";
-import { type HistoryPeriod, calculateDateRanges, formatLabel } from "@zoonk/utils/date-ranges";
-import { safeAsync } from "@zoonk/utils/error";
 import {
   aggregateScoreByMonth,
   aggregateScoreByWeek,
   aggregateScoreByYear,
-} from "@zoonk/utils/score-aggregation";
+} from "@zoonk/utils/aggregation";
+import { formatLabel } from "@zoonk/utils/chart";
+import { type HistoryPeriod, calculateDateRanges } from "@zoonk/utils/date-ranges";
+import { safeAsync } from "@zoonk/utils/error";
 import { cache } from "react";
 
 export type ScorePeriod = HistoryPeriod;
