+++ packages/core/src/audio/generate-language-audio.ts
@@ -1,7 +1,7 @@
 import "server-only";
 import { generateLanguageAudio as generateAudio } from "@zoonk/ai/tasks/audio";
-import { type TTSVoice } from "@zoonk/utils/constants";
 import { type SafeReturn } from "@zoonk/utils/error";
+import { type TTSVoice } from "@zoonk/utils/languages";
 import { toSlug } from "@zoonk/utils/string";
 import { uploadAudio } from "./upload-audio";
