+++ packages/utils/src/aggregation.test.ts
@@ -1,11 +1,89 @@
 import { describe, expect, it } from "vitest";
 import {
   type ScoredRow,
+  aggregateByMonth,
+  aggregateByWeek,
+  aggregateByYear,
   aggregateScoreByMonth,
   aggregateScoreByWeek,
   aggregateScoreByYear,
   findBestByScore,
-} from "./score-aggregation";
+} from "./aggregation";
+
+describe(aggregateByWeek, () => {
+  const dataPoints = [
+    { date: new Date(2026, 2, 2), value: 10 }, // Monday
+    { date: new Date(2026, 2, 3), value: 20 }, // Tuesday (same week)
+    { date: new Date(2026, 2, 9), value: 30 }, // Next Monday (different week)
+  ];
+
+  it("aggregates by sum", () => {
+    const result = aggregateByWeek(dataPoints, (point) => point.value, "sum");
+    expect(result.map((row) => row.value)).toEqual([30, 30]); // 10+20, 30
+  });
+
+  it("aggregates by average", () => {
+    const result = aggregateByWeek(dataPoints, (point) => point.value, "average");
+    expect(result.map((row) => row.value)).toEqual([15, 30]); // (10+20)/2, 30/1
+  });
+
+  it("returns sorted results", () => {
+    const reversed = [...dataPoints].toReversed();
+    const result = aggregateByWeek(reversed, (point) => point.value, "sum");
+    const times = result.map((row) => row.date.getTime());
+    expect(times).toEqual([...times].toSorted((left, right) => left - right));
+  });
+});
+
+describe(aggregateByMonth, () => {
+  const dataPoints = [
+    { date: new Date(2026, 0, 5), value: 10 },
+    { date: new Date(2026, 0, 20), value: 20 },
+    { date: new Date(2026, 1, 10), value: 30 },
+  ];
+
+  it("aggregates by sum", () => {
+    const result = aggregateByMonth(dataPoints, (point) => point.value, "sum");
+    expect(result.map((row) => row.value)).toEqual([30, 30]); // Jan: 10+20, Feb: 30
+  });
+
+  it("aggregates by average", () => {
+    const result = aggregateByMonth(dataPoints, (point) => point.value, "average");
+    expect(result.map((row) => row.value)).toEqual([15, 30]); // Jan: (10+20)/2, Feb: 30/1
+  });
+
+  it("returns sorted results", () => {
+    const reversed = [...dataPoints].toReversed();
+    const result = aggregateByMonth(reversed, (point) => point.value, "sum");
+    const times = result.map((row) => row.date.getTime());
+    expect(times).toEqual([...times].toSorted((left, right) => left - right));
+  });
+});
+
+describe(aggregateByYear, () => {
+  const dataPoints = [
+    { date: new Date(2025, 3, 10), value: 10 },
+    { date: new Date(2025, 8, 20), value: 20 },
+    { date: new Date(2026, 1, 5), value: 30 },
+  ];
+
+  it("aggregates by sum", () => {
+    const result = aggregateByYear(dataPoints, (point) => point.value, "sum");
+    expect(result.map((row) => row.value)).toEqual([30, 30]); // 2025: 10+20, 2026: 30
+  });
+
+  it("aggregates by average", () => {
+    const result = aggregateByYear(dataPoints, (point) => point.value, "average");
+    expect(result.map((row) => row.value)).toEqual([15, 30]); // 2025: (10+20)/2, 2026: 30/1
+  });
+
+  it("returns sorted results", () => {
+    const reversed = [...dataPoints].toReversed();
+    const result = aggregateByYear(reversed, (point) => point.value, "sum");
+    const times = result.map((row) => row.date.getTime());
+    expect(times).toEqual([...times].toSorted((left, right) => left - right));
+  });
+});
 
 function calcScore(correct: number, incorrect: number) {
   return (correct / (correct + incorrect)) * 100;
