+++ packages/utils/src/date-ranges.test.ts
@@ -1,11 +1,6 @@
 import { describe, expect, it, vi } from "vitest";
 import {
-  aggregateByMonth,
-  aggregateByWeek,
-  aggregateByYear,
-  buildChartData,
   calculateDateRanges,
-  formatLabel,
   formatPeriodLabel,
   getDefaultStartDate,
   validatePeriod,
@@ -151,108 +146,6 @@ describe(calculateDateRanges, () => {
   });
 });
 
-describe(formatLabel, () => {
-  it("formats month period as day + short month", () => {
-    const date = new Date(2026, 2, 15);
-    const result = formatLabel(date, "month", "en");
-    expect(result).toContain("Mar");
-    expect(result).toContain("15");
-  });
-
-  it("formats 6months period as week number", () => {
-    const date = new Date(2026, 0, 15);
-    const result = formatLabel(date, "6months", "en");
-    expect(result).toMatch(/^W\d+$/);
-  });
-
-  it("formats year period as short month", () => {
-    const date = new Date(2026, 2, 15);
-    const result = formatLabel(date, "year", "en");
-    expect(result).toBe("Mar");
-  });
-
-  it("formats all period as year string", () => {
-    const date = new Date(2026, 0, 1);
-    const result = formatLabel(date, "all", "en");
-    expect(result).toBe("2026");
-  });
-});
-
-describe(aggregateByWeek, () => {
-  const dataPoints = [
-    { date: new Date(2026, 2, 2), value: 10 }, // Monday
-    { date: new Date(2026, 2, 3), value: 20 }, // Tuesday (same week)
-    { date: new Date(2026, 2, 9), value: 30 }, // Next Monday (different week)
-  ];
-
-  it("aggregates by sum", () => {
-    const result = aggregateByWeek(dataPoints, (point) => point.value, "sum");
-    expect(result.map((row) => row.value)).toEqual([30, 30]); // 10+20, 30
-  });
-
-  it("aggregates by average", () => {
-    const result = aggregateByWeek(dataPoints, (point) => point.value, "average");
-    expect(result.map((row) => row.value)).toEqual([15, 30]); // (10+20)/2, 30/1
-  });
-
-  it("returns sorted results", () => {
-    const reversed = [...dataPoints].toReversed();
-    const result = aggregateByWeek(reversed, (point) => point.value, "sum");
-    const times = result.map((row) => row.date.getTime());
-    expect(times).toEqual([...times].toSorted((left, right) => left - right));
-  });
-});
-
-describe(aggregateByMonth, () => {
-  const dataPoints = [
-    { date: new Date(2026, 0, 5), value: 10 },
-    { date: new Date(2026, 0, 20), value: 20 },
-    { date: new Date(2026, 1, 10), value: 30 },
-  ];
-
-  it("aggregates by sum", () => {
-    const result = aggregateByMonth(dataPoints, (point) => point.value, "sum");
-    expect(result.map((row) => row.value)).toEqual([30, 30]); // Jan: 10+20, Feb: 30
-  });
-
-  it("aggregates by average", () => {
-    const result = aggregateByMonth(dataPoints, (point) => point.value, "average");
-    expect(result.map((row) => row.value)).toEqual([15, 30]); // Jan: (10+20)/2, Feb: 30/1
-  });
-
-  it("returns sorted results", () => {
-    const reversed = [...dataPoints].toReversed();
-    const result = aggregateByMonth(reversed, (point) => point.value, "sum");
-    const times = result.map((row) => row.date.getTime());
-    expect(times).toEqual([...times].toSorted((left, right) => left - right));
-  });
-});
-
-describe(aggregateByYear, () => {
-  const dataPoints = [
-    { date: new Date(2025, 3, 10), value: 10 },
-    { date: new Date(2025, 8, 20), value: 20 },
-    { date: new Date(2026, 1, 5), value: 30 },
-  ];
-
-  it("aggregates by sum", () => {
-    const result = aggregateByYear(dataPoints, (point) => point.value, "sum");
-    expect(result.map((row) => row.value)).toEqual([30, 30]); // 2025: 10+20, 2026: 30
-  });
-
-  it("aggregates by average", () => {
-    const result = aggregateByYear(dataPoints, (point) => point.value, "average");
-    expect(result.map((row) => row.value)).toEqual([15, 30]); // 2025: (10+20)/2, 2026: 30/1
-  });
-
-  it("returns sorted results", () => {
-    const reversed = [...dataPoints].toReversed();
-    const result = aggregateByYear(reversed, (point) => point.value, "sum");
-    const times = result.map((row) => row.date.getTime());
-    expect(times).toEqual([...times].toSorted((left, right) => left - right));
-  });
-});
-
 describe(getDefaultStartDate, () => {
   it("returns parsed date from ISO string", () => {
     const iso = "2026-01-15T00:00:00.000Z";
@@ -271,61 +164,6 @@ describe(getDefaultStartDate, () => {
     expect(result.toISOString().slice(0, 10)).toBe(expected.toISOString().slice(0, 10));
 
     vi.useRealTimers();
-  });
-});
-
-describe(buildChartData, () => {
-  const rawPoints = [
-    { count: 10, date: new Date(2026, 0, 5) },
-    { count: 20, date: new Date(2026, 0, 6) },
-    { count: 30, date: new Date(2026, 0, 20) },
-    { count: 40, date: new Date(2026, 1, 10) },
-  ];
-
-  it("returns daily data points for 'month' period (no aggregation)", () => {
-    const result = buildChartData(rawPoints, "month", "en");
-    expect(result.dataPoints).toHaveLength(4);
-    expect(result.dataPoints.map((dp) => dp.value)).toEqual([10, 20, 30, 40]);
-  });
-
-  it("aggregates to weekly sums for '6months' period", () => {
-    const result = buildChartData(rawPoints, "6months", "en");
-    expect(result.dataPoints.length).toBeLessThan(4);
-    const totalValue = result.dataPoints.reduce((sum, dp) => sum + dp.value, 0);
-    expect(totalValue).toBe(100);
-  });
-
-  it("aggregates to monthly sums for 'year' period", () => {
-    const result = buildChartData(rawPoints, "year", "en");
-    expect(result.dataPoints).toHaveLength(2);
-    expect(result.dataPoints.map((dp) => dp.value)).toEqual([60, 40]);
-  });
-
-  it("aggregates to yearly sums for 'all' period", () => {
-    const crossYearPoints = [
-      { count: 10, date: new Date(2025, 3, 5) },
-      { count: 20, date: new Date(2025, 8, 6) },
-      { count: 30, date: new Date(2026, 1, 10) },
-    ];
-    const result = buildChartData(crossYearPoints, "all", "en");
-    expect(result.dataPoints).toHaveLength(2);
-    expect(result.dataPoints.map((dp) => dp.value)).toEqual([30, 30]);
-    expect(result.dataPoints.map((dp) => dp.label)).toEqual(["2025", "2026"]);
-  });
-
-  it("returns empty data points and zero average for empty input", () => {
-    const result = buildChartData([], "month", "en");
-    expect(result).toEqual({ average: 0, dataPoints: [] });
-  });
-
-  it("calculates average correctly", () => {
-    const result = buildChartData(rawPoints, "month", "en");
-    expect(result.average).toBe(Math.round((10 + 20 + 30 + 40) / 4));
-  });
-
-  it("calculates average correctly for aggregated data", () => {
-    const result = buildChartData(rawPoints, "year", "en");
-    expect(result.average).toBe(Math.round((60 + 40) / 2));
   });
 });
