+++ packages/utils/src/string.ts
@@ -1,29 +1,8 @@
 import { NAME_PLACEHOLDER } from "@zoonk/utils/constants";
 import slug from "slug";
 
-const NUMERIC_ID_PATTERN = /^\d+$/;
-
 export function removeAccents(str: string): string {
   return str.normalize("NFD").replaceAll(/[\u0300-\u036F]/g, "");
-}
-
-export function validateOffset(value?: unknown): number {
-  const parsed = Number(value);
-  return Number.isFinite(parsed) ? Math.max(0, Math.trunc(parsed)) : 0;
-}
-
-export function parseNumericId(value: string): number | null {
-  if (!NUMERIC_ID_PATTERN.test(value)) {
-    return null;
-  }
-  return Number.parseInt(value, 10);
-}
-
-export function parseBigIntId(value: string): bigint | null {
-  if (!NUMERIC_ID_PATTERN.test(value)) {
-    return null;
-  }
-  return BigInt(value);
 }
 
 export function normalizeString(str: string): string {
@@ -32,15 +11,6 @@ export function normalizeString(str: string): string {
 
 export function toSlug(str: string): string {
   return slug(str.trim());
-}
-
-/**
- * Formats a 0-indexed position as a 2-digit display string.
- * @example formatPosition(0) // "01"
- * @example formatPosition(9) // "10"
- */
-export function formatPosition(position: number): string {
-  return String(position + 1).padStart(2, "0");
 }
 
 export function emptyToNull(value?: string | null): string | null {
